weather and time separation of concerns
