+++ src/main/java/com/codeup/plantapp/models/Weather.java
@@ -20,7 +20,7 @@ public class Weather{
     private LocalDateTime sunsetDTG;
     private long cloudiness;
     private String cloudDesc;
-    private double windSpeed;
+    private long windSpeed;
 
     public double getTempAvg() {
         return tempAvg;
@@ -40,7 +40,7 @@ public class Weather{
     public String getCloudDesc() {
         return cloudDesc;
     }
-    public double getWindSpeed() {
+    public long getWindSpeed() {
         return windSpeed;
     }
 
@@ -48,101 +48,11 @@ public class Weather{
 //      Date date = new Date(); // Fri Jun 09 08:42:40 CDT 2023
 //      User user = new User(date, "username", "first_name", "last_name", "San Antonio", "email", "password");
 //      Weather usersLocalWeather = getWeather(user);
-    public static Weather getWeather(User user) {
-    String apiKey = Keys.getWeather();
-    String city = user.getCity();
-    try {
-        URL url = new URL("https://api.openweathermap.org/data/2.5/weather?q=" + city + "&appid=" + apiKey +
-                "&units=imperial");
-        HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-        conn.setRequestMethod("GET");
 
-        BufferedReader reader = new BufferedReader(new InputStreamReader(conn.getInputStream()));
-        StringBuilder response = new StringBuilder();
-        String line;
 
-        while ((line = reader.readLine()) != null) {
-            response.append(line);
-        }
-        reader.close();
-
-        // Parse JSON response
-        JSONParser parser = new JSONParser();
-        JSONObject jsonResponse = (JSONObject) parser.parse(response.toString());
-
-        JSONObject atmObject = (JSONObject) jsonResponse.get("main");
-        double tempAvg = (double) atmObject.get("temp");
-        long humidity = (long) atmObject.get("humidity");
-
-        JSONObject sunObject = (JSONObject) jsonResponse.get("sys");
-        long sunrise = (long) sunObject.get("sunrise") * 1000;
-        LocalDateTime sunriseDTG = convertTimestampToLocalDateTime(sunrise);
-        long sunset = (long) sunObject.get("sunset") * 1000;
-        LocalDateTime sunsetDTG = convertTimestampToLocalDateTime(sunset);
-
-        JSONObject cloudObject = (JSONObject) jsonResponse.get("clouds");
-        JSONArray cloudArray = (JSONArray) jsonResponse.get("weather");
-        JSONObject cloudDescObject = (JSONObject) cloudArray.get(0);
-        long cloudiness = (long) cloudObject.get("all");
-        String cloudDesc = (String) cloudDescObject.get("description");
-
-        JSONObject windObject = (JSONObject) jsonResponse.get("wind");
-        double windSpeed = (double) windObject.get("speed");
-
-        return new Weather(tempAvg, humidity, sunriseDTG, sunsetDTG, cloudiness, cloudDesc, windSpeed);
-    } catch (Exception e) {
-        e.printStackTrace();
-        return null;
-    }
-}
-    public static LocalDateTime convertTimestampToLocalDateTime(long timestamp) {
-        Instant instant = Instant.ofEpochMilli(timestamp);
-        return LocalDateTime.ofInstant(instant, ZoneId.systemDefault());
-    }
-
-//      String weekDay = Weather.getWeekday(usersLocalWeather.getSunset()); // FRIDAY
-//      String month = Weather.getMonth(usersLocalWeather.getSunset()); // JUNE
-//      int day = Weather.getDay(usersLocalWeather.getSunset()); // 9
-//      int hour = Weather.getHour(usersLocalWeather.getSunset()); // 8
-//      int minute = Weather.getMinute(usersLocalWeather.getSunset()); // 32
-
-    public static String getWeekday(LocalDateTime dateTime) {
-        return dateTime.getDayOfWeek().toString();
-    }
-
-    public static String getMonth(LocalDateTime dateTime) {
-        return dateTime.getMonth().toString();
-    }
-
-    public static int getDay(LocalDateTime dateTime) {
-        return dateTime.getDayOfMonth();
-    }
-
-//  To make calculations by hour
-    public static int getHour(LocalDateTime dateTime) {
-        if (dateTime.getHour() > 12) {
-            return dateTime.getHour() - 12;
-        } else {
-            return dateTime.getHour();
-        }
-    }
-//  To make calculations by minute
-    public static int getMinute(LocalDateTime dateTime) {
-        return dateTime.getMinute();
-    }
-//  To print time for USER view
-    public static String printTime(LocalDateTime dateTime) {
-        int hour = dateTime.getHour();
-        int minutes = dateTime.getMinute();
-        if (hour > 12) {
-            return (hour - 12) + " : " + minutes + " PM";
-        } else {
-            return hour + " : " + minutes + " AM";
-        }
-    }
 
     public Weather(double tempAvg, long humidity, LocalDateTime sunrise, LocalDateTime sunset, long cloudiness, String cloudDesc,
-                   double windSpeed) {
+                   long windSpeed) {
         this.tempAvg = tempAvg;
         this.humidity = humidity;
         this.sunriseDTG = sunrise;
@@ -152,4 +62,5 @@ public class Weather{
         this.windSpeed = windSpeed;
     }
     public Weather () {}
+
 }
